Attach escalations to their parent objects
It makes absolutely no sense to keep them in a skiplist and access
them with complicated accessor-functions when we know we'll only
ever be interested in them when we're already looking at a host or a
service.

With this patch we cut the number of skiplists down to the object types
that actually have to be indexed, which is good. We can also simplify
a lot of algorithms and reduce runtime when we hit problems in networks
where escalations are configured, which is even better. Less memory
usage and less cpu spent. Win, win and win.

Signed-off-by: Andreas Ericsson <ae@op5.se>